+++ ecotrack-sjma/services/service-iot/src/services/cacheService.js
@@ -1,5 +1,5 @@
 const redis = require('redis');
-const logger = require('./src/utils/logger');
+const logger = require('../utils/logger');
 
 class CacheService {
   constructor() {